requirements.txt
Django>=1.9,<1.10
psycopg2>=2.6
django-debug-toolbar>=1.4
requests
beautifulsoup4
python-dateutil
html5lib
